+++ src/config.ts
@@ -155,6 +155,7 @@ export interface OptionConfig {
 export interface ExtensionConfigurationSettings {
     autoSelectActiveFolder: boolean;
     defaultActiveFolder: string | null;
+    exclude: string[];
     cmakePath: string;
     buildDirectory: string;
     installPrefix: string | null;
@@ -322,6 +323,10 @@ export class ConfigurationReader implements vscode.Disposable {
     get defaultActiveFolder(): string | null {
         return this.configData.defaultActiveFolder;
     }
+    get exclude(): string[] {
+        return this.configData.exclude;
+    }
+
     buildDirectory(multiProject: boolean, workspaceFolder?: vscode.ConfigurationScope): string {
         if (multiProject && this.isDefaultValue('buildDirectory', workspaceFolder)) {
             return '${sourceDirectory}/build';
@@ -596,6 +601,7 @@ export class ConfigurationReader implements vscode.Disposable {
     private readonly emitters: EmittersOf<ExtensionConfigurationSettings> = {
         autoSelectActiveFolder: new vscode.EventEmitter<boolean>(),
         defaultActiveFolder: new vscode.EventEmitter<string | null>(),
+        exclude: new vscode.EventEmitter<string[]>(),
         cmakePath: new vscode.EventEmitter<string>(),
         buildDirectory: new vscode.EventEmitter<string>(),
         installPrefix: new vscode.EventEmitter<string | null>(),